moved topic page pane
